Extracted Receipt related code
